abi-compliance-checker: Fix for Linuxbrew
Only delete abi-compliance-checker.cmd on OS X, since it is only created
if the OS has "win" in its name (i.e. Windows and Darwin).

Closes Linuxbrew/linuxbrew#1084

Closes Linuxbrew/homebrew-core#23.

+++ Formula/abi-compliance-checker.rb
@@ -16,7 +16,7 @@ class AbiComplianceChecker < Formula
 
   def install
     system "perl", "Makefile.pl", "-install", "--prefix=#{prefix}"
-    rm bin/"abi-compliance-checker.cmd"
+    rm bin/"abi-compliance-checker.cmd" if OS.mac?
   end
 
   test do
